HRONBOARIDNG/hr-assistant/backend/requirements.txt
# FastAPI and server dependencies
# If installation fails, try: pip install -r requirements.txt --no-build-isolation
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.8.0

# Document parsing
pdfplumber==0.10.2
python-docx==1.0.1

# Vector database options
chromadb==0.4.18

# Text processing and embeddings
langchain==0.0.335
openai==1.3.5
numpy

# Environment variables
python-dotenv==1.0.0